+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-08-06 22:49:57
-// Commit: fix(perf): correct layout constraints and overflow in perf
+// Updated at: 2026-08-08 12:25:26
+// Commit: feat(perf): add bi-directional event bus handler for perf
 
 class perfComponentState {
-  final String commitRef = '3434';
-  final String lastMessage = 'fix(perf): correct layout constraints and overflow in perf';
-  final DateTime timestamp = DateTime.parse('2026-08-06T22:49:57+05:00');
+  final String commitRef = '3460';
+  final String lastMessage = 'feat(perf): add bi-directional event bus handler for perf';
+  final DateTime timestamp = DateTime.parse('2026-08-08T12:25:26+05:00');
 }